requirements.txt
rich
time
random
pyyaml
getpass
hashlib
pyperclip